Convert sequences to ES Modules
