+++ lib/sequences/counter.mjs
@@ -6,11 +6,9 @@
  * @param {number} [step=1] The value to increment by each iteration. This may be positive or negative.
  * @yields numbers in the sequence
  */
-function* counter(start = 0, step = 1) {
-	while (true) {
-		yield start;
-		start += step;
-	}
+export default function* counter(start = 0, step = 1) {
+    while (true) {
+        yield start;
+        start += step;
+    }
 }
-
-module.exports = counter;
