+++ lib/sequences/index.mjs
@@ -1,15 +1,15 @@
-export const { iterate, safeIterate, reverse } = require("../base-iterators");
-export const combination = require("./combination");
-export const counter = require("./counter");
-export const random = require("./random");
-export const range = require("./range");
-export const repeat = require("./repeat");
-export const permute = require("./permute");
-export const permutation = require("./permutation");
-export const series = require("./series");
-export const shuffle = require("./shuffle");
-export const symbols = require("./symbols");
-export const thueMorse = require("./thueMorse");
-export const collatz = require("./collatz");
-export const fibonacci = require("./fibonacci");
-export const fizzBuzz = require("./fizzBuzz");
+export { iterate, safeIterate, reverse } from "../base-iterators/index";
+export { default as combination } from "./combination";
+export { default as counter } from "./counter";
+export { default as random } from "./random";
+export { default as range } from "./range";
+export { default as repeat } from "./repeat";
+export { default as permute } from "./permute";
+export { default as permutation } from "./permutation";
+export { default as series } from "./series";
+export { default as shuffle } from "./shuffle";
+export { default as symbols } from "./symbols";
+export { default as thueMorse } from "./thueMorse";
+export { default as collatz } from "./collatz";
+export { default as fibonacci } from "./fibonacci";
+export { default as fizzBuzz } from "./fizzBuzz";
