+++ lib/sequences/range.mjs
@@ -1,5 +1,5 @@
-const { takeWhile } = require("../chainables");
-const counter = require("./counter");
+import { takeWhile } from "../chainables/index";
+import counter from "./counter";
 
 /**
  * Generate a range of numbers progressing from `start` up to, but not including, `end`.
@@ -12,20 +12,18 @@ const counter = require("./counter");
  * @param {number} [step=1] The value to increment or decrement by.
  * @yields numbers in the sequence
  */
-function* range(
-	start = 0,
-	end = (() => {
-		let end = start;
-		start = 0;
-		return end;
-	})(),
-	step = (() => (start < end ? 1 : -1))()
+export default function* range(
+    start = 0,
+    end = (() => {
+        let end = start;
+        start = 0;
+        return end;
+    })(),
+    step = (() => (start < end ? 1 : -1))()
 ) {
-	let count = counter(start, step);
+    let count = counter(start, step);
 
-	let predicate = Math.sign(step) < 0 ? v => v > end : v => v < end;
+    let predicate = Math.sign(step) < 0 ? v => v > end : v => v < end;
 
-	yield* takeWhile(count, predicate);
+    yield* takeWhile(count, predicate);
 }
-
-module.exports = range;
