+++ lib/sequences/repeat.mjs
@@ -8,21 +8,19 @@
  * @param {Iterable|function} generatorFn
  * @param {...*} [args]
  */
-function* repeat(generatorFn, ...args) {
-	while (true) {
-		let it =
-			typeof generatorFn === "function"
-				? generatorFn(...args)
-				: generatorFn;
-		let empty = true;
+export default function* repeat(generatorFn, ...args) {
+    while (true) {
+        let it =
+            typeof generatorFn === "function"
+                ? generatorFn(...args)
+                : generatorFn;
+        let empty = true;
 
-		for (let value of it) {
-			empty = false;
-			yield value;
-		}
-		// Guard against empty iterators
-		if (empty) return;
-	}
+        for (let value of it) {
+            empty = false;
+            yield value;
+        }
+        // Guard against empty iterators
+        if (empty) return;
+    }
 }
-
-module.exports = repeat;
