--- lib/sequences/symbols.js
@@ -1,7 +1,0 @@
-function* symbols(name) {
-	while (true) {
-		yield Symbol(name);
-	}
-}
-
-module.exports = symbols;
