+++ lib/sequences/symbols.mjs
@@ -1,0 +1,5 @@
+export default function* symbols(name) {
+    while (true) {
+        yield Symbol(name);
+    }
+}
